requirements.txt
streamlit>=1.32.0
langchain>=0.1.17
langchain-openai>=0.1.6
openai>=1.16.1
python-dotenv>=1.0.1